Move LicenseMapping to the spdx package as SpdxLicenseMapping
As it is all about SPDX licenses.
